+++ src/features/createPost/CreatePostModalContent.css
@@ -1,0 +1,176 @@
+.modal-header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 16px 24px;
+}
+.modal-title {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  font-size: 24px;
+  line-height: 32px;
+  letter-spacing: -0.6px;
+
+  .lucide-sparkles {
+    color: var(--color-azure-61);
+  }
+}
+.password-guide {
+  font-size: 12px;
+  line-height: 16px;
+  color: #6b7280;
+  margin-top: 8px;
+}
+
+.modal-close {
+  position: absolute;
+  top: 17px;
+  right: 17px;
+  cursor: pointer;
+}
+
+/* 본문 */
+.modal-body {
+  padding: 0 24px 16px;
+  overflow-y: auto;
+}
+.modal-row {
+  display: flex;
+  gap: 12px;
+  margin-bottom: 16px;
+}
+.modal-section {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+}
+.modal-section label {
+  font-size: 18px;
+  margin-bottom: 8px;
+}
+.modal-section input,
+.modal-section textarea,
+.modal-section select {
+  width: 100%;
+  padding: 10px;
+  font-size: 14px;
+  border: 1px solid #ccc;
+  border-radius: 6px;
+}
+.message-input {
+  min-height: 200px;
+  resize: none;
+}
+
+.modal-actions {
+  display: flex;
+  justify-content: flex-end;
+  gap: 10px;
+}
+
+/* 이모지 */
+.emoji-panel {
+  margin: 12px 24px;
+}
+.emoji-grid {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(40px, 1fr));
+  gap: 10px;
+}
+.emoji-img {
+  width: 40px;
+  height: 40px;
+  cursor: pointer;
+  border-radius: 6px;
+  transition: transform 0.2s;
+}
+.emoji-img:hover {
+  transform: scale(1.1);
+}
+
+/* 푸터 */
+.modal-footer {
+  padding: 16px 24px;
+  display: flex;
+  align-items: center;
+  justify-content: flex-end;
+  gap: 10px;
+}
+.error-message {
+  color: #e00;
+  font-size: 13px;
+  margin-right: auto;
+}
+.modal-actions button {
+  padding: 10px 16px;
+  font-size: 14px;
+  border: none;
+  border-radius: 6px;
+  cursor: pointer;
+}
+.modal-actions button:first-of-type {
+  background: #ccc;
+}
+.modal-actions button:last-of-type {
+  background: #4a90e2;
+  color: #fff;
+}
+
+.emoji-toggle {
+  margin-top: 8px;
+  padding: 10px 13px;
+  border-radius: 10px;
+  color: #4a90e2;
+  border: 1px solid var(--color-azure-88);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 4px;
+  float: right;
+}
+
+.emoji-panel {
+  background: #f9f9f9;
+  border-radius: 12px;
+  padding: 16px;
+  margin: 12px 0;
+  box-shadow: 0 2px 6px rgba(0, 0, 0, 0.05);
+}
+
+.emoji-header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 8px;
+}
+
+.emoji-grid {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(40px, 1fr));
+  gap: 10px;
+}
+
+.emoji-img {
+  width: 40px;
+  height: 40px;
+  object-fit: contain;
+  cursor: pointer;
+  border-radius: 6px;
+  transition: transform 0.2s ease;
+}
+
+.emoji-img:hover {
+  transform: scale(1.1);
+}
+
+.selected-emoji {
+  margin-top: 8px;
+}
+
+/* 반응형 */
+@media (max-width: 768px) {
+  .modal-row {
+    flex-direction: column;
+  }
+}
