--- src/pages/MessageModal.css
@@ -1,285 +1,0 @@
-/* 디자인 시스템 적용할 것! 지금은 임시 스타일 */
-
-.modal-backdrop {
-  position: fixed;
-  inset: 0;
-  background: rgba(0, 0, 0, 0.5);
-  z-index: 1000;
-
-  display: flex; /* flex 컨테이너로 설정 */
-  justify-content: center; /* 수평 정렬 */
-  align-items: center; /* 수직 정렬 */
-}
-
-.modal {
-  background: #e6f2ff;
-  width: 90%;
-  max-width: 768px;
-  padding: 25px 0;
-  border-radius: 12px;
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.2);
-
-  display: flex;
-  flex-direction: column;
-  max-height: 90vh;
-  overflow: hidden;
-}
-
-.modal-title {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  font-size: 24px;
-  line-height: 32px;
-  letter-spacing: -0.6px;
-}
-
-.password-guide {
-  font-size: 12px;
-  line-height: 16px;
-  color: #6b7280;
-  margin-top: 8px;
-}
-
-.modal-close {
-  position: absolute;
-  top: 17px;
-  right: 17px;
-  cursor: pointer;
-}
-
-.lucide-sparkles {
-  color: var(--color-azure-61);
-}
-
-.modal-row {
-  display: flex;
-  gap: 12px;
-}
-
-.modal-section {
-  flex: 1;
-  margin-bottom: 16px;
-  display: flex;
-  flex-direction: column;
-}
-
-.modal-section label {
-  font-size: 18px;
-  line-height: 28px;
-  margin-bottom: 8px;
-}
-
-.modal-section input,
-select {
-  background-color: #e6f2ff;
-  &:focus {
-    outline: none;
-    border-color: #4a90e2;
-  }
-}
-
-.modal-section input,
-.modal-section textarea,
-.modal-section select {
-  width: 100%;
-  padding: 10px;
-  border: 1px solid #ccc;
-  border-radius: 6px;
-  font-size: 14px;
-}
-
-.modal-section textarea {
-  border: 2px solid var(--color-azure-88);
-  background-color: #fff;
-  resize: none;
-  &:focus {
-    outline: none;
-    border-color: #4a90e2;
-  }
-}
-
-.modal-header,
-.modal-footer {
-  position: relative;
-  padding: 0 25px;
-  flex-shrink: 0;
-}
-
-.modal-body {
-  flex-grow: 1;
-  padding: 25px;
-  overflow-y: auto;
-}
-
-.modal-actions {
-  display: flex;
-  justify-content: flex-end;
-  gap: 12px;
-}
-
-.modal-section textarea {
-  min-height: 254px;
-}
-
-.char-count {
-  text-align: right;
-  font-size: 12px;
-  color: gray;
-  margin-top: 4px;
-}
-
-.error-message {
-  color: red;
-  margin-bottom: 12px;
-  font-size: 13px;
-}
-
-.modal-actions {
-  display: flex;
-  justify-content: flex-end;
-  gap: 10px;
-}
-
-.modal-actions button {
-  padding: 10px 16px;
-  font-size: 14px;
-  border: none;
-  border-radius: 6px;
-  cursor: pointer;
-}
-
-.modal-actions button:first-of-type {
-  background-color: #ccc;
-}
-
-.modal-actions button:last-of-type {
-  background-color: #4a90e2;
-  color: white;
-}
-
-@media (max-width: 768px) {
-  .modal-backdrop {
-    align-items: flex-start;
-  }
-
-  .modal {
-    width: 100vw;
-    margin: 0;
-    max-width: none;
-    overflow-y: auto;
-  }
-}
-
-@media (max-width: 390px) {
-  .modal-row {
-    flex-direction: column;
-  }
-}
-
-.emoji-button-wrapper {
-  text-align: right;
-  margin-top: 8px;
-}
-
-.emoji-toggle {
-  margin-top: 8px;
-  background: none;
-  border: none;
-  color: #4a90e2;
-  cursor: pointer;
-  font-size: 14px;
-  padding: 4px 0;
-}
-
-.emoji-panel {
-  background: #f9f9f9;
-  border-radius: 12px;
-  padding: 16px;
-  margin: 12px 0;
-  box-shadow: 0 2px 6px rgba(0, 0, 0, 0.05);
-}
-
-.emoji-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 8px;
-}
-
-.emoji-grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(40px, 1fr));
-  gap: 10px;
-}
-
-.emoji-img {
-  width: 40px;
-  height: 40px;
-  object-fit: contain;
-  cursor: pointer;
-  border-radius: 6px;
-  transition: transform 0.2s ease;
-}
-
-.emoji-img:hover {
-  transform: scale(1.1);
-}
-
-.selected-emoji {
-  margin-top: 8px;
-}
-
-.emoji-toggle {
-  padding: 10px 13px;
-  border-radius: 10px;
-  border: 1px solid var(--color-azure-88);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 4px;
-  float: right;
-}
-
-//--------
-
-
-.message-canvas {
-  position: relative;
-  height: 300px;
-  border: 1px solid #ddd;
-  border-radius: 8px;
-  overflow: hidden;
-  padding: 8px;
-  background: #fdfdfd;
-}
-
-.message-input {
-  width: 100%;
-  height: 100%;
-  resize: none;
-  border: none;
-  padding: 10px;
-  font-size: 14px;
-  background: transparent;
-  z-index: 0;
-  position: relative;
-}
-
-.modal-header,
-.modal-footer {
-  padding: 16px 24px;
-  flex-shrink: 0;
-}
-
-.modal-body {
-  overflow-y: auto;
-  padding: 0 24px 16px;
-  flex-grow: 1;
-}
-
-.modal {
-  display: flex;
-  flex-direction: column;
-  max-height: 90vh;
-}
